fix: paginator and null cursor values
This patch fixed the paginator when cursor values can be nil. We
assume that the order by will always follow postgres defaults, which
are:

  * asc: asc_nulls_last
  * desc: desc_nulls_first

Which means that other options are not supported, namely,
:asc_nulls_first and :desc_nulls_last.

Previously, the code would remove from the where clause all fields
which the cursor value is nil. While this works for pagination using a
single field, it breaks if you use two or more fields.

This patch looks at the value instead and modifies the where
expression accordingly - depending on the whether the value is null or
not.